+++ src/main/java/roomescape/reservation/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.reservation;
 
 import java.util.ArrayList;
 import java.util.List;
